+++ clease/calculator/__init__.py
@@ -1,9 +1,11 @@
+from .calc_cache import CleaseCacheCalculator
 from .clease import Clease, MovedIgnoredAtomError
 from .clease_vol_dep import CleaseVolDep
 from .util import attach_calculator, get_ce_energy
 
 __all__ = [
     "Clease",
+    "CleaseCacheCalculator",
     "MovedIgnoredAtomError",
     "attach_calculator",
     "get_ce_energy",