+++ clease/montecarlo/observers/lowest_energy_observer.py
@@ -1,6 +1,7 @@
 import logging
 import numpy as np
-from ase.calculators.singlepoint import SinglePointCalculator
+import ase
+from clease.calculator import CleaseCacheCalculator
 from .mc_observer import MCObserver
 
 logger = logging.getLogger(__name__)
@@ -12,24 +13,50 @@ class LowestEnergyStructure(MCObserver):
     atoms: Atoms object
         Atoms object used in Monte Carlo
 
+    track_cf: bool
+        Whether to keep a copy of the correlation functions for the
+        emin structure. If enabled, this will be stored in the ``lowest_energy_cf``
+        variable. Defaults to False.
+
     verbose: bool
         If `True`, progress messages will be printed
     """
 
     name = "LowestEnergyStructure"
 
-    def __init__(self, atoms, verbose=False):
+    def __init__(self, atoms: ase.Atoms, track_cf: bool = False, verbose: bool = False):
         super().__init__()
         self.atoms = atoms
-        self.lowest_energy = np.inf
-        self.lowest_energy_cf = None
-        self.emin_atoms = None
+        self.track_cf = track_cf
         self.verbose = verbose
-
-    def reset(self):
-        self.lowest_energy = np.inf
+        # Silence pylint
         self.lowest_energy_cf = None
-        self.emin_atoms = None
+        self.reset()
+
+    def reset(self) -> None:
+        self.lowest_energy_cf = None
+        self.emin_atoms: ase.Atoms = self.atoms.copy()
+        # Keep a reference directly to the calculator cache
+        self._calc_cache = CleaseCacheCalculator()
+        self.emin_atoms.calc = self._calc_cache
+        self.lowest_energy = np.inf
+
+    @property
+    def lowest_energy(self) -> float:
+        return self.emin_results["energy"]
+
+    @lowest_energy.setter
+    def lowest_energy(self, value: float) -> None:
+        """The currently best observed energy"""
+        self.emin_results["energy"] = value
+        # Keep a copy without needing to look up the calculator
+        # all the time, for quicker lookup.
+        self._lowest_energy_cache = value
+
+    @property
+    def emin_results(self) -> dict:
+        """The results dictionary of the lowest energy atoms"""
+        return self._calc_cache.results
 
     @property
     def calc(self):
@@ -37,9 +64,10 @@ class LowestEnergyStructure(MCObserver):
 
     @property
     def energy(self):
+        """The energy of the current atoms object (not the emin energy)"""
         return self.calc.results["energy"]
 
-    def __call__(self, system_changes):
+    def __call__(self, system_changes) -> None:
         """
         Check if the current state has lower energy and store the current
         state if it has a lower energy than the previous state.
@@ -49,20 +77,19 @@ class LowestEnergyStructure(MCObserver):
             `clease.montecarlo.observers.MCObserver`
         """
 
-        if self.emin_atoms is None or self.energy < self.lowest_energy:
+        if self.energy < self._lowest_energy_cache:
             self._update_emin()
             dE = self.energy - self.lowest_energy
-            msg = "Found new low energy structure. "
-            msg += f"New energy: {self.lowest_energy} eV. Change: {dE} eV"
-            logger.info(msg)
+            msg = "Found new low energy structure. New energy: %.3f eV. Change: %.3f eV"
+            logger.info(msg, self.lowest_energy, dE)
             if self.verbose:
-                print(msg)
+                print(msg % (self.lowest_energy, dE))
 
-    def _update_emin(self):
-        self.lowest_energy_cf = self.calc.get_cf()
+    def _update_emin(self) -> None:
+        if self.track_cf:
+            self.lowest_energy_cf = self.calc.get_cf()
         self.lowest_energy = self.energy
 
-        # Store emin atoms, and attach a cache calculator
-        self.emin_atoms = self.atoms.copy()
-        calc_cache = SinglePointCalculator(self.emin_atoms, energy=self.energy)
-        self.emin_atoms.calc = calc_cache
+        # Avoid copying the atoms object multiple times. We can only ever have
+        # changed the symbols, so copy those over from the atoms object.
+        self.emin_atoms.numbers = self.atoms.numbers