+++ tests/montecarlo/test_obs.py
@@ -1,4 +1,3 @@
-from multiprocessing.sharedctypes import Value
 import pytest
 import ase
 import numpy as np